feat: create database connection
Created a file and class to handle the PostgreSQL database
connections.  This uses a database connection pool to attempt following
best practices and to allow future use of multiple clients.

Having one connection and keeping it alive, as done in previous
projects, works fine when there is a single connected client.  When
having multiple clients or making frequent queries, a single connection
becomes a bottleneck.

Only integration or end-to-end tests would be needed, which will be
done later when there are API endpoints.