feat: implement User update and delete methods
Note that update does not currently return anything, because the only
data that can be changed is the password and passwords should not be
returned.

Note that delete is called on an User instance and does not remove the
username instance variable.  User instances would need to be
deleted/removed/unassociated after calling this method.