test: use toBe instead of toEqual for expects
In Jest expect statements, toEqual's main purpose is to compare object
instances recursively.  Even though it can compare primitives just
fine, it is not the proper matcher to be used.

The expects that compare String values have been changed to use the
toBe matcher, because toBe is the one that should be used for comparing
primitives.