fix: update database schema for documents table
The documents table allowed for duplicate document names.  Earlier,
this was allowed, because there was not a foreseen need to have unique
names.  The only possible way was to retrieve a list of all documents
belonging to a user, then use the ID to then retrieve a specific
document.  This is fine for the majority of use cases when interacting
thru the frontend, but not directly thru the API.

In the documents table, a document's name is now required to be unique
for each owner.  This helps reduce confusion when looking thru a list
of documents, as well as allow for finding a specific document by name
instead of ID.