feat: add helpers
Added convertPropsForSqlUpdate function, which takes an Object of
properties that will be used to update a database table entry.  This
returns the SQL substring that will be used after the SET command and
an Array of values that will be passed to the database query method
thru the query config Object.

Added camelToSnakeCase function, which converts a String in camel case
to snake case.