feat: implement Document model
Implemented the model with add, getAll, get, and update methods.

Template literals are indented towards the left, because it allows the
String to be readable and consistent in both the code and logs.
Otherwise, the logs will show significant indentations.

Other solutions are no good.  Using backslashes (\) at the end of
String lines do not remove the extra spacing in logs.  Using Arrays and
joining the Strings together makes it less readable in the code and
might require extra processing.