refactor: include more error handling for db query
In PostgresDb.query, added more parameters for log prefix and callback.

The log prefix will be used for the logging, within query, to name the
specific functions that uses query.  This helps to see what classes and
methods are accessing the database, instead of simply stating a query
is being executed, which was not helpful.

The callback is used if additional, specific error handling is needed
from a calling function.

document.js and user.js have been refactored accordingly.