fix: make section names unique
Originally, the concept was to allow users to create any number of
sections and call them anything.  Then it was to only have a preset
amount of sections, because there were only a handful number of them
for resumes.  I allowed duplicate section names, because it might have
been useful.  Now, I foresee that I would never need duplicate section
names.

The database schema has been updated to now have section names be
unique.  The sections table is still used, rather than referring to
sections by static names, because in the future, it allows other info
to be added for each section.