refactor: generalize tests for models of contents
Moved common tests and test structure into _testCommon.js.  This should
help speed up testing the other data models, such as sections,
educations, experiences, and skills.  The contact info model has been
excluded, because it is too different to easily include in the
generalization.

Only one function, runCommonTests, needs to be called to run the common
tests.  It takes an Object that contains the various information needed
to run tests for a specific model.  Understanding the configuration
needed is extremely vital to getting the tests to run.

In the afterAll block for the test suite, the sections table is now
truncated as well, because it does not get truncated during the
cascaded truncation of users.

For parameterized tests within runCommonTests, an extra parameter,
expected, is added, because computing the expected data within the
generalized tests might not be possible.

Certain sets of tests, like getAll and get, can be disabled.  This is
done, because certain models might not work with these tests.
Originally, it was because the contact info model was going to be
included in the generalization.  This might not be needed anymore and
will need to be reviewed later.

In the test for successful updating, a specific line for the Document
model is included to expect the last updated property as any date.
This is needed, because that property is only set by the app; client
input does not include it.  Other similar lines might be included for
other models.

In Document and Section classes, the static property of the SQL
statement, used to help convert database table attribute names to
JavaScript names, has been renamed to "_allDbColsAsJs".  This is done
to help with generalization.

Moved test data into _testData.js.

Note that it takes a while to clean up after the tests are done.  Jest
will warn about not exiting after one second.

runCommonTests should work for the following models:
Document
Section
Education
Experience
Skill
Certification
Project
TextSnippet