fix: use time zone in database schema
While developing the TextSnippet data model, inserting and retrieving
the timestamp into and from the database caused incorrect time zone
conversions.  Even though everything was done in UTC 0, timestamps
retrieved from the database were still adjusted according to the local
time zone.

Instead of trying to spend extra time figuring out how to force
everything into UTC 0, timestamp now includes the time zone and
node-postgres and PostgreSQL will automatically do the conversions.  So
TIMESTAMP is changed to TIMESTAMPTZ.

There was also an issue with not being able to find a specific text
snippet by ID and version.  This was due to PostgreSQL storing
timestamps down to microseconds and JavaScript only being able to work
down to milliseconds.  In other words, searches were performed using
timestamps with 3 decimal places, but the text snippet entries in the
database had versions to 6 decimal places, thus not being able to find
any entries from the database.

This was solved by setting the precision for timestamps in the database
schema to 3.