test: set up tests for TextSnippet
Modified _testCommon.js to allow TextSnippet to utilize it.  The
modifications involved including the second primary key of
TextSnippet.  In the tests for update, certain expects for dates had to
be for any date, since dates are created with the current date.

Added test data for text snippets for common use in _testData.js.

Added test data for tests specific to text snippets in
_textSnippetTestData.js.  This was normally included with the test
file, but having additional specific tests for the TextSnippet data
model does not allow that.

The tests specifically for TextSnippet could not easily be included in
the generalized tests in _testCommon.js.  As a result, they had to be
written in another file.  The specific tests could not be in
textSnippet.test.js, because running specific and common tests together
causes the database pool of clients to shutdown when one set of tests
finish.

With two test files for TextSnippet, the test data specifically for
TextSnippet had to be imported for textSnippetSpecific.test.js.  If the
test data was in textSnippet.test.js, that would cause the
runCommonTests function to be executed when the require keyword imports
it.

Consequently, test data had to be moved into its own file,
_textSnippetTestData.js, to avoid running any functions.