feat: implement TextSnippet class
Implemented the add, getAll, get, update, and delete methods.

Note that the update method creates a new database entry, instead of
updating an existing one.

Note that the delete method causes all current instances of newer
versions of a text snippet to be obsolete.  There is no way to set all
the instances' parent property to null in the TextSnippet class.  This
would have to be done outside, in the service level.  The rest of the
methods are not impacted by this, because they do not utilize the
parent property.  Do note that when retrieving the parent of a text
snippet, that parent may not exist.

Specific tests for TextSnippet were created in a different file,
separated from the file that runs the common tests, as explained in the
commit "test: set up tests for TextSnippet".  In summary, it is because
running both common and specific tests together causes the database
pool of clients to shutdown after one set of tests finishes.