feat: implement Skill class
Implemented add, getAll, get, update, and delete methods.

Added common test data for skills in _testData.js.

The common/generalized tests could not be used, because a skill table
entry has attributes with references to existing text snippets.  The
text snippets needed to be created first, and then the primary keys, ID
and version, could then be saved.  With this way, a lot of
asynchronicity is required, and consequently, there were some code that
could not be easily put into the common tests to make it happen.

I realize now that I do not need to let the app and database create the
ID and version of text snippets used for test data.  The
common/generalized could have been used.  In the future, consider
refactoring and using static values for ID and version.

Also note that a new test was added for the add method.  It tests for
when new skills are created with owners and text snippets that don't
exist in the database.