chore: remove owner parameter for get in models
In the data models' get methods, the owner parameter was not necessary,
as the id parameter is enough to find a specific table record, because
it is the primary key.  Having the SQL statement with an additional
WHERE clause for owner is redundant.

For Document and Skill classes, the error message, for not finding a
table record, has been updated to also include both the ID and name,
for better clarification.