refactor: remove extra variable in test setup
In document_x_section and document_x_education tests, sectionData and
educationData are removed, and their code logic are now placed in the
beforeAll setup.  They only existed to help create test data and could
be integrated along with the code that inserted the data in the
sections and educations tables.