fix: put unique constraints on relationship tables
documents_x_experiences and documents_x_projects used their own ids for
primary keys, so the two table ids in each of them no longer had the
SQL UNIQUE constraint.  This can cause multiples of the same
experiences or projects to exist in a document.

UNIQUE constraints have now been placed on the document ID and the
other table ID for each of the relationship tables.  This will now only
allow an experience or project to only show up once in a document.