feat: implement Document_X_Skill
Implemented add, getAll, get, and delete.

Included documents_x_skills test data in _testData.js.

There is no update method, because documents_x_skills table does not
currently have the position attribute, so there is nothing to update.
If changing the skill of a table entry, the entry should be deleted
instead and a new one should be created.