feat: implement user registration route
Added new Express route in app.js for user authentication.

Added a JSON schema for input data given during user registration.
This restricts characters to a length range, so that stored data do not
overflow String/text size limit.  A regex pattern is used for the
password, which checks for a number, uppercase letter, lowercase
letter, and symbol.

Implemented the "/api/v1/auth/register" route, which handles user
registration.  At this moment, a service layer is not necessary, but
this needs to be reviewed again later.

Updated _testData.js with valid passwords for user accounts, since
there is a JSON schema now.

Note that there are tests that tests the JSON schema.

Also added a new npm script, testDebug, in package.json to make
debugging easier by extending the Jest timeout.