feat: implement user sign-in route
Added a JSON schema for input data given during user sign-in.  This
restricts characters to a length range to be consistent with the
registration schema.

Implemented the "/api/v1/auth/signin" route, which handles user
sign-in.  At this moment, a service layer is not necessary, but this
needs to be reviewed again later.

In auth.test.js, the outside beforeEach code for truncating the users
table is moved into the describe block for register, because only the
register route needs to have a cleared table before each test.  The
sign-in route reuses the singly created user.