refactor: use variables in User model
Added a variable to store the table name.
Added a variable for columns that will be returned from a SQL query.

Also extracted all returned attributes, except password, from a SQL
query, instead of simply directly inputting needed parameters when
creating a new User instance.  This will help in the future if more
data is stored in a user account.  Tests are updated accordingly.