feat: create static/class User.update
When working on the users route for updating account data, there were
no existing user instances.  This would have then required making an
additional database call to fetch user data, put it into an instance,
and the update method could then be used.

To forego this double database call, the User update method has now
been split into a class and instance method.

The update methods also now return User instances.  Currently, they
only contain username, which results in no difference with the previous
implementation, but this was done in anticipation for future attributes
that could be included.  For the class version, a User instance is
returned instead of a generic Object to be consistent with the other
class methods.

Note that the test for the instance update method is now a unit test,
because the instance method is small and isolated.  There is no need to
make it an integration test and complicate things.