feat: implement user account update route
Added new Express route in app.js for user operations.

Added a JSON schema for input data given during user account update.
Note that it uses the same regex for the new password as during
registration.  Note that at least one piece of data is required, but
currently only consists of password.  This requirement was done in
consideration for possible future additional data.

Implemented the "PATCH /api/v1/users/:username" route, which handles
updating user account info.  At this moment, a service layer is not
necessary, but this needs to be reviewed again later.  There is some
logic at the bottom that considers data other than password, which may
be added in the future.