feat: implement contact info PUT route
Added a new JSON schema for the inputs of contact information.  There
are no required properties, so the check for the full name, when
creating a new database entry, has to be done within the route code.

Implemented the route that will either create a new database entry for
contact info or update it.

For the "PUT /users/:username/contact-info" route, 2 database calls
were made: a read operation followed by update or create.  This logic
was used, because the alternative requires a create followed by a read
and update, which equates to 1 or 3 database calls.  ContactInfo
currently only has the update method for an instance, that is why a
read operation is needed if doing the alternative.  Refactoring and
adding a static update method could be done later to reduce database
calls to at most 2, if there's time.

In the test file, the beforeAll block in "PATCH /users/:username" is
moved to an outer scope, because the new "PUT
/users/:username/contact-info" route requires authentication tokens as
well.  Note that all users are registered, and tokens from all users
are taken.