fix: update JSON schemas for email and username
For the contact info JSON schema, the email property was not allowing
empty strings.  Users need to be able to give empty strings so that
their emails can be cleared out.

Previously, usernames were allowed to have special characters and
whitespace.  At the time, there wasn't a use case to not allow this.
Now, potential issues could arise if they are used in other places,
like in URLs.  Usernames are now restricted to letters and numbers to
avoid special handling, where ever they are used.