feat: implement documents POST and GET routes
Implemented:
POST /users/:username/documents
GET /users/:username/documents

The POST route does not allow creating a master resume, because that
should be created on user registration and should never be deleted.

The GET route gets document info and not content for all documents
belonging to a user.  This will be utilized when displaying all
documents to a user.

The GET operation to retrieve a document's contents should be done
later, as the tests could utilize other API POST endpoints to create
the contents.

Also added the documents route to app.js.

Also created a JSON schema for new document inputs.