feat: create static/class Document.update
For at least some of the API endpoint routes, there would not be any
existing model instances to call update on.  This would have required
sending two database calls: one to retrieve the data and another to
update it.

Since the client would be sending the ID(s) of the relevant entity, one
database call could be made instead, by creating an additional class
update method.

The current tests still pass, but the tests need to be modified and
more tests need to be added, just like the User model.  Since Document
tests uses the common test suite, the suite can not be changed now and
will need to be changed after all models have been updated.