feat: implement update document route
This route only updates the meta-data of a document (name, is template,
is locked) and not the content.

Created the JSON schema for the inputs for updating documents.

In documents.test.js, truncating the documents table before each test
is moved to outside all describes/suites, because it was common to all
describes.