chore: have User.update throw NotFoundError
User's static/class update method was throwing AppServerError when the
provided username was not found.  This is incorrect for the class
update method, as the username would be provided by the client.

The class update method now throws a NotFoundError, and the instance
update method now rethrows the error as AppServerError.  If the
instance method could not find the user, then it's because the instance
is holding onto a deleted user, which would be a developer's mistake.

In the new test, "Rethrows NotFoundError as AppServerError.", the
instance update method is run twice for the assertion section, because
it was simpler to write out and the code for the instance update method
is small enough that the impact of running twice is insignificant.