feat: implement document delete service
Since there was some complex logic for the document delete route, it
was abstracted/moved to a service-level file.

In deleteDocument, the document to delete is retrieved first and
verified if it belongs to the user that wants to delete it.  This is
done to prevent users from deleting anyone's documents.

To have the delete operation be idempotent, no Errors are thrown if the
document is not found; this means that the document is already deleted.