feat: implement document update service
Since there was some complex logic for the document update route, it
was abstracted/moved to a service-level file.

In updateDocument, the document to update is retrieved first and
verified if it belongs to the user that wants to update it.  This is
done to prevent users from deleting anyone's documents.