fix: prevent users from updating other users' docs
The PATCH /users/:username/documents/:docId route was allowing anybody
to make updates to anyone's documents.  A check for the correct
document owner was missing.

Instead of directly calling Document.update, documentService.js ->
updateDocument was created, in a previous commit, and is now used to
check if the user requesting the update to a document is the owner of
the document.  In other words, this commit simply has the route utilize
updateDocument.

A new test was added to avoid this issue.