chore: remove manual input of sample data IDs
The sample data that were being seeded into the production database
were having their IDs being manually set.  Originally, this was
supposed to help refer to dependent table data, such as when populating
relationship tables, as the IDs would be known and not generated by the
database.

The problem with this was that the ID sequences were not being
updated.  This causes any insertions that utilize the automatic ID
generation to cause a duplicate key violation, because the first ID
that will be generated would be 1.

Instead of fiddling with the sequence tables, seed.sql has now been
updated with the manual setting of IDs removed.  Any IDs of previously
inserted data can be deducted by viewing the seed file and looking at
the number of inserted records.