fix: correct app import in server.js and misc.
In server.js, the app import was not updated after the database code
was removed from app.js.  The export from app.js is no longer an
object; it is just app.

In middleware/auth.js, the ensureLoggedIn middleware was giving an
ambiguous log and no error message.

In routes/document.js, the PATCH /users/:username/documents/:docId
route's JSDoc's parameters are declared as optional, to help clarify
that not all of them are required.