feat: create master resume on user creation
Each user will need a master resume and thus should have a master
resume created for them.  The best time would be on user account
creation.

Created a service-level user function, createUser in userService.js, to
help abstract out the code that will register a user and create a
master resume document.

The user registration tests have been updated to also check the
existence of a master resume.

Unit tests for createUser were not needed as it only plugs in values
into functions.