feat: allow updating master resume
For the documents route tests, added tests related to updating master
resumes.  Also did a refactor by moving the common "updateData"
variable in tests out to an outer scope.

Added logic for handling a master resume in documentService.js ->
updateDocument.  The logic allows only the document name to be
changed.  Also added tests related to updating master resumes and
modified existing updateDocument tests to indicate if document is a
master resume.

The decision for only allowing the document name of master resumes to
be changed is based on:

1. Master resumes are supposed to be consistently changing and would
not benefit from being locked from changing.  Other resumes and
templates need to retrieve the content from master resumes, as the
others won't be able to make edits.

2. Master resumes are not supposed to be templates, as not all pieces
of content are supposed to be in a particular resume.