refactor: abstract out document owner validation
The code, for retrieving a specified document and checking if that
document's owner is the user updating, deleting, or accessing it, was
being repeated and going to be repeated.

Therefore, the code has been abstracted out into the
validateDocumentOwner function to reduce duplication of the code and of
tests.

documentService.test.js has been updated to use a mock of the
validateDocumentOwner function instead of the Document.get method.
"Throws an Error if document does not belong to user." tests have been
removed, because that test is now in serviceHelpers.test.js.

In documentService.test.js, for the updateDocument describe block, the
extra tests, for ensuring that Errors from validateDocumentOwner and
document.update continue to be passed, have been removed, because
Errors are normally assumed to be passed from called functions.  Having
those extra tests made reading the tests more difficult and are
unnecessary.

In documentService.test.js, for the deleteDocument describe block, the
test for passing along Errors from validateDocumentOwner has been kept,
because there is logic that stops passing the NotFoundError thrown by
validateDocumentOwner.  Conversely, the other test for passing Errors
from document.delete has been removed.