feat: implement document-section create route
Created the POST /users/:username/documents/:docId/sections/:sectionId
route.  This creates a document-section relationship.

A new schema, document_x_sectionNew.json, is created to more easily
check the URL parameters by having jsonschema handle that logic.  This
can reduce the amount of tests to write or can consolidate JSON
schema-related tests in one place.

The test setup in routes/sections.test.js is modified by setting up
user data, which allows sections to be attached to specific documents.

A service-level function, createDocument_x_section, is created in a new
file, sectionService.js.  This helps to contain the logic between
client input validation at the start of a route call and database
interaction at the end of a route call.

No unit tests are written for createDocument_x_section, because it only
plugs in values into functions.  The code is covered through the
integration tests on the route, in routes/sections.test.js.