feat: implement get all document's sections route
Implemented the route that will get all sections in a document.

Created a new JSON schema, in urlParams.json, that will help validate
the URL parameters.  This should be used for other routes as well and
should help abstract and reduce the code for validation.

Created a service-level function, getSections in
services/sectionService.js, for any code that is not input validation.
Currently, it only validates document ownership of the requesting
user.  Unit tests for this is skipped, because there is no logic; it is
only plugging in values into functions.  Code coverage is performed
through integration tests.

In the tests for the section routes in sections.test.js, in addition to
the integration tests for the new route, moved the "POST
/users/:username/documents/:docId/sections/:sectionId" URL to create
document-section relationships to the top and to an outer scope,
because it will be used in other describe blocks to set up tests.

Also in sections.test.js, for the test to get all sections of a
document in the correct order, the setup is an attempt to get the
sections in order of position, regardless of position value or
document_x_section insertion order.  Though currently, the test does
not set up the sections in the document with non-sequential position
values (ex. 0, 3, 9), because there is no logic to allow for that yet.