feat: order returned relationship data by position
For the relationship models, document_x_education,
document_x_experience, document_x_section, and
experience_x_text_snippet, the getAll method now returns the results in
order by position.  Using the results in this way should be the most
often use case.  This allows for content to be removed from a document,
leaving an empty position (ex. 1, 3), and yet still return content in
order without having to update all positions.

In regards to sections...

Number of sections for test data has been increased to 3.  This is done
to better test the different position ranges, though I suspect that
having more than 2 sections is not necessary.

In document_x_section.test.js, generalized the section insertion in
test setup in beforeAll block.  Previously, the number of sections
inserted was manually done and only inserted 2 sections.  Now, any
number of sections can be inserted.  The use of values in the
queryConfig and the use of parameterized SQL statements are not
necessary, as the values are test values created by the developer and
not given by clients.