chore: update document-section next position logic
When creating a new document-section relationship, the old logic simply
assumes all documents_x_sections' positions are in proper sequential
order.  Always keeping the positions correct required unnecessary
processing.

The new logic allows positions to be any value, and the only thing that
matters is the relativity between positions.  For example, the top
section of a document can have position 2, the second section of
position 10, and the third section of position 20.  The values don't
matter, only that the second section comes after the first and before
the third, and that the third section comes after the second.  This
allows for deleting relationships without having to update their
positions.