feat: defer database unique constraints
Applied the "DEFERRABLE INITIALLY DEFERRED" option to the unique
constraints in relationship tables.  This will defer checking the
constraint until a SQL transaction is done.

This was done so that the positions of relationship table records can
be easily updated.  There will be no need to set a record to an unused
position and then set it to the desired position.  All of the positions
can be updated in one go, creating conflicting positions, and being
sorted out in the end.