chore: remove getting client in PostgresDb.query
PostgresDb.query had some temporary code to connect to the pg Pool and
get a pg Client.  The temporary code was placed as a reminder to
retrieve a Client before executing query.

The code has now been removed, and in its stead, a pg Client should now
be passed into the query method.  If one is not supplied, the Pool is
used instead, which is good for one-statement (non-transaction)
queries.  A getClient method is also implemented as a result, so that a
Client from the Pool can be obtained.

Tests are skipped, because there isn't much logic.