feat: make section service to update positions
Created updateDocument_x_sectionPositions, which helps to update the
section positions in a document.  It ensures that the document being
interacted with belongs to the user requesting the update and that the
given section IDs are exactly the ones in the document.  In other
words, the number of given IDs must match the number of sections in the
document and must exist in the document.  Having the exact section IDs
helps to keep the positions consistent and sequential.  This method
also ensures that the sections to reposition belong to the user.