feat: make update section positions route
Implemented the route that allows updating the positions of all
sections in a document.  Note that all and only those sections in the
document must be given.  This is done to ensure consistency of the
positions, because there can be sections with skipped positions, due to
the ability to simply remove sections without needing to update all
section positions every time.

Created a new JSON schema, in schemas/document_x_sectionPositions.json,
that checks the request body with section ID inputs, to validate if
they are unique integers in an array.