test: add test for sections route
Added a test for checking that a new document-section relationship is
given the correct position when there are existing relationships but
the relationships are non-sequential and out of order.  This test is
for the use case of adding a relationship after sections have been
repositioned and with some sections removed from the document.