feat: implement create education route
Created a route that will create a new education and document-education
relationship entry in the database.  Currently, only the master resume
can have new educations placed in it, due to time constraints, but in
the future, this may change.

This route will set the position of the new education after the last
education's position.  Note that the test for this is written, but is
commented, because it needs some of the other routes for test setup.
Once the other routes are done, it can be uncommented.

Added the education routes file to the Express setup in app.js.

Created a JSON schema for the new education data.  Date patterns do
simple validations for date format YYYY-MM-DD and does not consider
leap years or shorter February dates.  The pattern is taken from
https://stackoverflow.com/a/43347172.  String length requirements are
arbitrary and can be changed in the future.  They are there to have
some sensible restrictions instead of none.  Thorough unit testing of
the schema will have to be done later, along with the other schemas.