feat: implement service to create doc-edu relation
Created createDocument_x_education service function for the route that
creates a document-education relationship.  This helps to validate
document ownership, get the next position an education can be in a
document, and attach the education at that position in the document.

The code is copied from the section service, but refactoring will have
to be done later as the shared code might be used for relationships
that have an extra ID property or do not have the position property.