test: refactor createDocument_x_...
createDocument_x_education and createDocument_x_section did not need to
have their own test data, so to maintain consistency and not cause the
tests to falsely pass if the data structure changes, the data from
_testData.js is used instead.

This change is possible because the functions under test do not
directly interact with the test data.  The test data is only handled by
the getLastPosition helper function.  The only logic in the functions
under test is to calculate the next position an education or section
can be.

The expected new position is now calculated within the tests, because
it should always be one value higher than the last position, so passing
that data into the tests was not necessary.

Also renamed the variable "mockObject" to remove ambiguity.