feat: implement document-education create route
Created the route that will attach an existing education to a document,
by creating the document-education relationship.

Updated the urlParams.json schema to include validation for
educationId, because the route uses it and it will be provided by the
client.

In the tests, educationsForRawClientInputs is moved to a higher scope,
because the new tests now uses it.

In the tests, for the POST
/users/:username/documents/:documentId/educations test suite, a new
afterAll block is added to clean up the education records in the
database created by the test suite.  The education records now
interfere with the tests for creating document-education
relationships.  Originally, it wasn't cleaned up after running the test
suite because there wasn't a need to and it was already going to be
cleaned up after the file is done running.

In _testCommon.js, added a new variable to hold the URL that contains
document and education IDs.