test: rename variables and reduce repeated code
Moved code, for getting the document ID of the master resume, out of
every test and into the beforeAll block to remove duplicate code.

Removed the expectedDocument_x_section variables.  They were used to
help clarify the expected values, but weren't necessary after all, and
the removal makes the tests more readable.

Changed the docId variable name to documentId to be consistent and
because it foregoes value assignment in Object creations.