feat: implement update education route
Created the route to update an education entry.  This only updates the
education info and not any relationships.  With that in mind, it does
require that the document the education is in to be the master resume.

Created the related schema in educationUpdate.json, which is adapted
from educationNew.json.  These can be refactored using references, but
will require research, and so will be skipped for now.

In the tests, only a small subset of the allowed properties to update
is used, because it was simpler.  In the future, there might be a need
to test with multiple updated properties, which would contain one
change, changes for all required properties, and changes for required
and optional properties.  This would require using parameterized tests
and original educations with and without optional properties.

In the test file, educations.test.js, the username variable is moved to
a parent scope to reduce duplication.  For the "POST
/users/:username/documents/:documentId/educations/:educationId" route,
in the beforeAll block, the document ID did not need to be retrieved
again, as that was already in the masterDocumentIds variable.  Other
changes in existing code are refactors to simplify the code.