chore: correct title in documentUpdate schema
The title for the document update schema was not updated when it was
copied over.  Also, extra lines are added and lines are moved around to
make the file more readable.

Document_X_Education.get method's JSDoc is also corrected.  When the
line was copied over, the type for the educationId parameter was
overlooked.  The education ID will always be a number, because it is
created by the database, even though its input type may be a String or
Number.