feat: implement route to reposition educations
Created the route, service, and model function to update the positions
of all educations within a document.  Also created a function in the
Education model to get all educations in a document, so that the route
returns the educations in their updated order.  Note that exactly all
educations need to be given to ensure consistency of the positions,
because there can be instances where the positions are not sequential.

Most of the code is similar to section and hasn't been refactored due
to time.

Unit and integration tests are skipped due to time, but basic
end-to-end testing has been done in Postman.