feat: implement delete doc-edu route
Created the route to delete a document-education relationship.

Created a service function to help verify document ownership.
Education verification is not necessary, because the education would
not be in the document if it didn't belong to the user, the route will
do nothing if the education is not in the document, and there is no
access to the education data.

Tests are skipped for now due to time, but basic functionality has been
tested in Postman.