fix: add verification for education ownership
When attaching an education to a document, a user was able to attach
another user's education to their own.  In other words, when creating a
document-education relationship, there was no check to see if the
education belonged to the user.

A check has now been added, which retrieves the education to check the
owner of it.

The tests have been updated in regards to the new call to the Education
model.  An additional test has been created to verify that the
createDocument_x_education function throws an Error.