chore: provide better errors when adding to doc
For creating document-section and document-education relationships, if
the relationship already exists, a database error is returned to the
client.  This is confusing as the error message is not user-friendly.

In sectionService.js -> createDocument_x_section and
educationService.js -> createDocument_x_education, added a check for
the database error with code 23505, which will occur if adding
duplicate document relationships, and a more descriptive re-thrown
Error, to clarify that a relationship already exists.

Added unit tests for each function accordingly as well.